implement count

+++ tracing/tracer.go
@@ -7,5 +7,5 @@ type Tracer interface {
 	// otherwise asynchronous
 	Trace(string, func(), *sync.WaitGroup) error
 
-	Count(bucket string, value uint64) error
+	Count(bucket string, value int) error
 }
